ci: narrow clang-tidy scope and fix lattice test widening

+++ tests/lattice_test.cpp
@@ -152,7 +152,8 @@ static void test_sublattice_readonly() {
 
 static void test_temp_lattice_memory() {
     // Small enough to stay in memory.
-    TempLattice<float> lat(IPosition{4, 4}, 1024 * 1024);
+    TempLattice<float> lat(IPosition{4, 4},
+                           static_cast<std::size_t>(1024) * 1024);
     CHECK(!lat.is_paged());
     CHECK(lat.is_writable());
